add a sensible mapping for GPerlFilename ('localized file name').
Fix podify_methods to take only the package name, and remove the second
my  in the same scope.

Minor re-work to logic of when to skip and when to count methods; intent was
to support hidden xsubs, but those are caught at the parser level now.
Changed the message that gets written when all methods are defined but missing.

svn path=/trunk/; revision=341