rework _INSTALL_OVERRIDES handling one more time.  rather than have a mess of inherited methods which chain up, Glib::Type::register_object looks for _INSTALL_OVERRIDES in all classes of the new type's ancestry, calling it if found.  this allows proper overriding behavior and lets us add vfunc implementations without requiring changes to all client code.
Glib::Type::register_object now *always* instantiates the class upon
registration, to avoid problems with it maybe or maybe not existing later.
we also leak the reference to avoid the class needing to be recreated,
since we can't properly recreate it later, and perl doesn't let classed
die in normal circumstances, anyway.

svn path=/trunk/; revision=492