Fix subclassing with {pspec, get, set}-style properties
Previously setting such a property defined by a superclass did not
reach those get/set funcs but instead the GET_PROPERTY/SET_PROPERTY
(in the superclass).  Or if the subclass had another get/set funcs
property (of a different name) at the same position N in the
properties list then the subclass funcs were reached instead of the
superclass ones.

https://bugzilla.gnome.org/show_bug.cgi?id=634350