Handle ownership of objects with floating refs correctly
Objects with floating refs (for example, GInitiallyUnowned descendants) are
created with a floating reference that needs to be "sunk".  Enforce this,
irregardless of the 'own' setting.

The custom converter for GtkObject in Gtk2 has always been enforcing own=1 and
installing a custom sink func, effectively provoking the exact same behavior as
with this commit.  Thus, this commit does not change the Gtk2::Object handling.