Handle ownership of objects with floating refs differently
The special handling of floating objects introduced by a previous commit
turned out to be too invasive, philosophically and practically.
Practically, because Gtk2/t/GtkAction.t started failing, and I did not
see a way to fix it.  And philosophically, because the only reason this
special handling was introduced was to account for some other special
handling in gobject-introspection.  But the reaction to this should live
in Glib::Object::Introspection, not in Glib.

So instead, we now just register a custom sink func for
GInitiallyUnowned which does ref_sink+unref.  Since sink funcs are
called only when we are supposed to own the object, this means that we
touch the floating flag only if we own the object.  This differs
slightly from the Gtk2::Object semantics: its wrappers always sink
floating refs (by enforcing own=TRUE), except when they are inside
GValues (signal invocation, for example).