Ensure timely destruction of initial wrapper of custom subclasses
Alter the init function for custom subclasses (the one which invokes the sub
INIT_INSTANCE) so that the initial SV wrapper is destroyed as soon as it is not
needed anymore, rather than when the next FREETMPS happens.  This makes the ref
counting more predictable but should otherwise not be observable.