Fix libpthread-related building issues on OpenBSD
On OpenBSD, any program that directly or indirectly wants to load libpthread.so
must do so from the start.  But when perl is built without ithreads, it will
also most likely not be compiled with "-pthread".  When libglib/libgobject then
go and try to load libpthread.so, the loader will error out.

Fix by setting LD_PRELOAD=libpthread.so at the relevant points in the build and
test process.