Sort the output of podify_signals in a deterministic way.
The output of Glib::Type->list_signals is unsorted. This results in the
generated pod files to be in random order which makes packages using signals
compile not reproducibly [1].

It was suggested to Glib upstream [2] to sort the output of g_signal_list_ids
but it got rejected. Thus the change has to be made here.

[1] https://wiki.debian.org/ReproducibleBuilds/About
[2] https://bugzilla.gnome.org/show_bug.cgi?id=743863

Author: akira <marivalenm@gmail.com>

https://bugzilla.gnome.org/show_bug.cgi?id=748796