Install constant numeric subs for all enum and flags values
Install, for example, a constant sub Glib::IOCondition::HUP returning
the number 16.

This is useful when a function, signal or property expects or returns an
enum or flags value, but the API specification is such that our special
enum and flags string handlers are not invoked.  Example:
Gtk3::TextTag's "weight" property accepts any positive integer but
predefined values are given by the Pango::Weight enum.