Glib::GenPod: supply default namespace blurbs
Namespace blurbs end up in the NAME section of POD and man pages.  In man
pages, they are thus used by 'whatis'.

If no blurb is specified for a namespace, try to figure out
which ctype corresponds to that (using regexps), check if they really
do (using the Glib::Type register stuff), and if they do add a whatis
that says "wrapper for $ctype".

https://bugzilla.gnome.org/show_bug.cgi?id=753468